refactor(tests): split test_selectors.py into a focused core_selectors package (#443)
The 1214-LOC tests/teatree_core/test_selectors.py was the largest
remaining oversized standalone test module. Split it verbatim by
selector concern into tests/teatree_core/core_selectors/ (package name
avoids shadowing the stdlib selectors module under pytest prepend
import):

- test_cache_and_format.py: cache primitives and pure formatting helpers
- test_queues.py: interactive/headless queues, stale-claim reaping,
  overlay scoping
- test_activity.py: active-session, recent-activity, last-result
- test_task_detail.py: build_task_detail
- test_pr_and_automation.py: PR action-required and automation summary

Pure relocation: no src/ change, no assertion or test-body change. All
19 test classes are AST-identical to the original; the same 69 tests
collect and pass. Only import lines were narrowed per file.

Relates-to #443
